requirements.txt
Flask>=2.3
httpx>=0.27
pytest>=8.0
python-dotenv>=1.0
